+++ modules/layers/src/text-layer/multi-icon-layer/multi-icon-layer-fragment.glsl.ts
@@ -19,7 +19,7 @@ out vec4 fragColor;
 void main(void) {
 #ifdef MODULE_COLLISION
   if (collision.visibilityPass) {
-    fragColor = vec4(collision_testBounds(collision_pickingColor), 0.0, 0.0, 1.0);
+    fragColor = collision_getBoundsData();
     return;
   }
 #endif